Feat: ItemController Validate, Authentication

+++ src/main/java/com/example/demo/web/ItemController.java
@@ -16,9 +16,13 @@ import org.springframework.data.web.PageableDefault;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
+import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import java.util.List;
+import java.util.Objects;
 
 @Controller
 @RequiredArgsConstructor
@@ -48,27 +52,44 @@ public class ItemController {
     }
 
     @PostMapping("/new")
-    public String addItem(@ModelAttribute ItemAddDto itemAddDto, Authentication authentication) {
+    public String addItem(@Validated @ModelAttribute(name = "item") ItemAddDto itemAddDto, BindingResult bindingResult, Authentication authentication) {
+        if (bindingResult.hasErrors()) {
+            log.info("errors={}", bindingResult);
+            return "items/addItemForm";
+        }
         log.info("itemAddDto ={} ", itemAddDto);
         itemService.saveItem(itemAddDto, authentication.getName());
         return "redirect:/";
     }
 
     @GetMapping("{id}/edit")
-    public String updateItemForm(@PathVariable Long id, Model model) {
+    public String updateItemForm(@PathVariable Long id, Model model, Authentication authentication, RedirectAttributes redirectAttributes) {
         ItemGetDto item = itemService.getItem(id);
+        if (!Objects.equals(item.getMemberName(), authentication.getName())) {
+            redirectAttributes.addFlashAttribute("flag", true);
+            return "redirect:/items/{id}";
+        }
         model.addAttribute("item", item);
         return "items/edit";
     }
 
     @PostMapping("{id}/edit")
-    public String updateItem(@PathVariable Long id, @ModelAttribute ItemUpdateDto itemUpdateDto) {
+    public String updateItem(@PathVariable Long id, @Validated @ModelAttribute(name = "item") ItemUpdateDto itemUpdateDto, BindingResult bindingResult) {
+        if (bindingResult.hasErrors()) {
+            log.info("error={}",bindingResult);
+            return "items/edit";
+        }
         itemService.updateItem(id,itemUpdateDto);
         return "redirect:/items";
     }
 
     @PostMapping("/{id}/delete")
-    public String deleteItem(@PathVariable Long id) {
+    public String deleteItem(@PathVariable Long id, Authentication authentication,RedirectAttributes redirectAttributes) {
+        ItemGetDto item = itemService.getItem(id);
+        if (!Objects.equals(item.getMemberName(), authentication.getName())) {
+            redirectAttributes.addFlashAttribute("flag", true);
+            return "redirect:/items/{id}";
+        }
         itemService.delete(id);
         return "redirect:/items";
     }
